# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: OrbitOS API
  description: |
    REST API for OrbitOS - An AI-native operating system for designing, running, and transforming businesses.

    ## Authentication
    All endpoints except `/health` require JWT authentication via Bearer token.

    ## Multi-Tenancy
    All organization-scoped resources are accessed via `/organizations/{orgId}/...` paths.
    The authenticated user must have appropriate membership in the organization.

    ## SOC 2 Compliance
    This API implements SOC 2 Type II controls. See SEC-001 for details.
  version: 0.8.0
  contact:
    name: OrbitOS Team
    email: api@orbitos.io
  license:
    name: Proprietary
    url: https://orbitos.io/license
servers:
  - url: http://localhost:5000/api/v1
    description: Local development
  - url: https://api.orbitos.io/v1
    description: Production
...
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Authenticate user
      operationId: login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      operationId: getCurrentUser
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserWithMemberships'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============================================================================
  # ORGANIZATIONS (ENT-003)
  # ============================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /auth/login
...
  responses:
...
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "NOT_FOUND"
        message:
          type: string
          example: "Resource not found"
        details:
          type: object
          additionalProperties: true
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
...
    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        expiresIn:
          type: integer
          description: Seconds until expiration
        user:
          $ref: '#/components/schemas/User'

    # ============================================================================
    # USER SCHEMAS (ENT-001)
    # ============================================================================
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        avatarUrl:
          type: string
          format: uri
          nullable: true
        phone:
          type: string
          nullable: true
        timezone:
          type: string
          example: "America/New_York"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    UserWithMemberships:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            memberships:
              type: array
              items:
                $ref: '#/components/schemas/MembershipSummary'

    # ============================================================================
    # ORGANIZATION SCHEMAS (ENT-003)
    # ============================================================================
...
    AccessLevel:
      type: string
      enum: [owner, admin, member, viewer]
      description: |
        - owner: Full control, can delete organization
        - admin: Full control except delete org and demote owners
        - member: Can create/edit resources, cannot manage members
        - viewer: Read-only access
...
    MembershipSummary:
      type: object
      properties:
        organizationId:
          type: string
          format: uuid
        organizationName:
          type: string
        organizationSlug:
          type: string
        accessLevel:
          $ref: '#/components/schemas/AccessLevel'
...
security:
  - bearerAuth: []